Return empty list instead of raising NotFoundError
In most cases where previously a NotFoundError was raised, now simply an empty list is returned. This means it is not left to the user to catch an exception simply when no results are returned.

A NotFoundError is still raised if when attempting to construct a Compound from a CID using `Compound.from_cid(cid)` if the CID does not exists.