Make get_properties accept underscore style property names
The attributes on Compound objects are underscore separated, consistent with Python conventions. However this means they are inconsistent with the property names required by the PUG REST API. This commit allows underscore separated names to be used with the `get_properties` function, in addition to the previously supported CamelCase names.