Update gettingstarted.rst
Re-run the instructions with PubChemPy (version 1.0.4) in Python 3.9.2, gentle revision.

Beside the more obvious change about `print foo` to `print(foo)`,  the PubChem database
apparently underwent some consolidation.  As a result some resulting lists do not contain
any more this many entries as in the elder documentation; perhaps with the exception of
synonyms known to the database.