update print instructions in the documentation
By the previous commmit, the documentation still included the
old Python2 print statement.  It has been superseeded by Python3's
print function this commit adopts.  On occasion, the results
querring the PubChem database differ from the earlier version of
the documentation which are considered, too.

A check run of the functionality described by the documentation
(except for code in advanced.rst, not tested) with Python 3.9.2,
pandas 1.3.1, jupyter-notebook 6.2.0, PubChemPy 1.0.4 in Linux
Debian 11/bullseye, branch testing, was successful.