ci(conftest.py): provide a rate limiting fixture
The PubChem database limits the number of queries per unit of
time.  The threshold appears to be lower, than the GitHub
runners normally launch pytest checks which eventually may
yield to a refusal by PubChem's interface, and hence a failure
of a pytest check.  This is now addressed with a throttle
on pytest's side.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>