Further refinement of new SMILES API (#115)
* Add smiles property to Compound

Just as an alias of the absolute_smiles, which is almost always what is desired.

* Deprecation warnings for old SMILES properties

canonical_smiles and isomeric_smiles are no longer returned by the PubChem API.

* Add new SMILES names to property map

* Better deprecated docstring

* Remove absolute_smiles in favor of smiles

Simplify by only having a smiles property for the absolute isomeric SMILES. Also ensure that the deprecated canonical_smiles and isomeric_smiles are not included when converting a Compound to a dictionary.

* Update docs for SMILES changes

* Update notebooks for SMILES changes

* Update tests for SMILES changes